requirements.txt
flask
redis
PyPDF2
sentence-transformers
openai
python-dotenv
torch --index-url https://download.pytorch.org/whl/cpu
